Version 10.0.0-beta
##### Enhancements

This beta version release includes major changes with several improvements and optimizations :

- Enhanced the coding standards of SDK to follow industry programming styles and best practices.
- Enhanced security standards of SDK.
- Reduced code between the business layer and persistence layer for optimization of SDK performance.
- Added internal parameter validations in the API function.
- ApiKey and ApiSecret usage redundancy removed.
- All LoginRadius related features need to be defined once only and SDK will handle them automatically.
- Improved the naming conventions of API functions for better readability.
- Better Error and Exception Handling for LoginRadius API Response in SDK.
- Revamped complete SDK and restructured it with latest API function names and parameters.
- Added detailed description to API functions and parameters for better understanding.
- Updated the demo according to latest SDK changes.
- Implemented API Region Feature.

##### Added new multiple APIs for better user experience

- Update Phone ID by UID
- Upsert Email
- Role Context profile
- MFA Resend OTP
- User Registration By Captcha
- Get Access Token via Linkedin Token
- Get Access Token By Foursquare Access Token
- Get Active Session By Account Id
- Get Active Session By Profile Id

##### Removed APIs:

- GetCompanies API

+++ Samples/dot-net-demo/dot-net-demo/wwwroot/js/profileControl.js
@@ -225,9 +225,9 @@ $("#btn-user-getcustomobj").click(function () {
             '<tr>').appendTo("#table-customobj > tbody:last-child");
 
         for (let i = 0; i < ret.data.length; i++) {
-            $("<tr><td>" + ret.data[i].id + "</td></tr>").appendTo("#table-customobj > tbody:last-child");
+            $("<tr><td>" + ret.data[i].Id + "</td></tr>").appendTo("#table-customobj > tbody:last-child");
             $("<td>", {
-                text: JSON.stringify(ret.data[i].customObject)
+                text: JSON.stringify(ret.data[i].CustomObject)
             }).appendTo("#table-customobj > tbody:last-child > tr:last-child");
         }
     });
@@ -277,7 +277,12 @@ $( "#btn-user-createrole" ).click(function() {
     });
 });
 
-$( "#btn-user-deleterole" ).click(function() {
+$("#btn-user-deleterole").click(function () {
+    if ($("#user-roles-deleterole").val() === "") {
+        $("#user-deleterole-message").text("Please input a valid Role.");
+        $("#user-deleterole-message").attr("class", "error-message");
+        return;
+    }
     $.ajax({
         method: "DELETE",
         url: m_options.deleteRoleUrl + "?role=" + $("#user-roles-deleterole").val(),
@@ -334,17 +339,20 @@ let profileUpdate = function() {
         }
     }).done(function (ret) {
         console.log(ret);
-        if (ret.fullName === null) {
+        if (ret.FullName === null) {
             $("#profile-name").html("");
         } else {
-            $("#profile-name").html("<b>" + ret.fullName + "</b>");
+            $("#profile-name").html("<b>" + ret.FullName + "</b>");
         }
-        $("#profile-provider").text("Provider: " + ret.provider);
-        $("#profile-email").text(ret.email[0].value);
-        $("#profile-lastlogin").text("Last Login Date: " + ret.lastLoginDate);
-        $("#user-updateaccount-firstname").val(ret.firstName);
-        $("#user-updateaccount-lastname").val(ret.lastName);
-        $("#user-updateaccount-about").val(ret.about);
+        if (ret.ImageUrl != null && ret.ImageUrl != undefined && ret.ImageUrl != "") {
+            $("#profile-image").attr('src', ret.ImageUrl)
+        }
+        $("#profile-provider").text("Provider: " + ret.Provider);
+        $("#profile-email").text(ret.Email[0].Value);
+        $("#profile-lastlogin").text("Last Login Date: " + ret.LastLoginDate);
+        $("#user-updateaccount-firstname").val(ret.FirstName);
+        $("#user-updateaccount-lastname").val(ret.LastName);
+        $("#user-updateaccount-about").val(ret.About);
     });
 }
 
@@ -361,7 +369,7 @@ let roleUpdate = function() {
         for (let i = 0; i < ret.data.length; i++) {
             $("<tr></tr>").appendTo("#table-allroles > tbody:last-child");
             $("<td>", {
-                text: ret.data[i].name
+                text: ret.data[i].Name
             }).appendTo('#table-allroles > tbody:last-child > tr:last-child');
         }
     });
@@ -376,11 +384,11 @@ let roleUpdate = function() {
         }
     }).done(function(ret) {
         $('#table-userroles tr:not(:first)').remove();
-        if (ret && ret.roles) {
-            for (let i = 0; i < ret.roles.length; i++) {
+        if (ret && ret.Roles) {
+            for (let i = 0; i < ret.Roles.length; i++) {
                 $("<tr></tr>").appendTo("#table-userroles > tbody:last-child");
                 $("<td>", {
-                    text: ret.roles[i]
+                    text: ret.Roles[i]
                 }).appendTo('#table-userroles > tbody:last-child > tr:last-child');
             }
         }
